refactor: modularize storage logic by splitting `storage.js` into `idb-store.js` and `board-serializer.js`, enhancing maintainability and clarity
